Refactor: collection패키지 제거 후 Lotto클래스 domain으로 이동
