+++ src/main/java/lotto/domain/User.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.collection.Lotto;
-
 import java.util.List;
 
 public class User {
